Investigate lack of colors and add debug info to item.c
Cleaned-up file of trailing spaces and leading tabs.
Changed several '#if 0' to use '#ifdef DEBUG' which can now be turned-on
as an option in ./configure.

Lack of colors can be due to XAllocColorCells() not working as expected.
XAllocColorCells() was created at an earlier time with older video cards
such as monochrome or possibly similar.